# #2317 Reduce workspace popup loading time by 98% when there are many files